refactor: move logic to trade store
